requirement.txt
numpy==1.16.0
opencv-python==3.4.5.20
pyautogui
system
wget
pandas
matplotlib
tensorflow-gpu==1.14.0
pafy
youtube-dl
python-vlc
